requirements.txt
streamlit
pyyaml
plotly
pandas
